+++ components/auth/form-warning/form-warning.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {View, Text} from 'react-native';
 
-import {WarningStyles} from './form-warning.styles';
+import {WarningStyles} from './styles';
 
 export const FormWarning = props => (
     <View style={WarningStyles.errorContainer}>
