+++ components/auth/sign-up/styles.js
