+++ components/common/search/search.js
@@ -3,8 +3,8 @@ import {TextInput, View} from 'react-native';
 import {FontAwesomeIcon} from '@fortawesome/react-native-fontawesome';
 import {faSearch} from '@fortawesome/free-solid-svg-icons';
 
-import {SearchStyles} from './search.style';
-import {BaseStyles} from '../../app.styles';
+import {SearchStyles} from './styles';
+import {BaseStyles} from '../../../app.styles';
 
 export class Search extends Component {
     constructor() {
