+++ components/counter/styles.js
