+++ components/main/group/styles.js
