+++ components/main/styles.js
